Refactor code, split classes into different files
